Fixes to the framelimiter to make it more accurate.
This change improves the accuracy of the framelimiter so it yields the expected result. Before this change, it was impossible to actually achieve the default maximum framerate of 120 FPS.

Please note that the Windows code does not have a sleep function with a high enough granularity to take advantage of the improved precision, so Windows builds will still have an imprecise FPS cap.