fix a bug where the ai buys 1 unneeded harvester
when it buys a refinery it gets a free harvester, but it succeeded to
buy another one just before the free harvester appears. It shows when
capturing an ai refinery with a harvester in it and then selling the
refinery. The ai will rebuild the refinery asasp, and was buying the
unneeded refinery as soon as it appeared. Just added a delay based on
JustBuilt, which is reseted when the free harvester is given.