shows the power in a tooltip
as does the remastered version when the mouse hovers the power area.
It's a kind of hack since the help tip function insists on having an int
for the text and here it's dynamic so it can't be a static translation,
and you'll get the word Power in all languages, oh well... !
It shows Power: Power/Drain where Power and Drain are ints.