some ways out of funpark
once a funpark has been activated, 2 ways to get back to normal game is
to select a misson either from the new missons dialog or from the
campaign missions dialog. Notice that it can't be done for "new game"
because it would conflict with the "funpark" on the command line...
funpark is still a hack and it shows here. To reset "new game" you have
to quit !