... and let's try imgui !!!
This is still a test, but a promising one.
The only place this is used is for the "new missions" dialog, it
replaces both this one and the old campaign mission one at the same
time... It's quite fun to see a hires dialog on top of a lores game
screen !
Quite a big commit, it includes the imgui files used because the project
seems to encourage people to merge the used files by their projects, so
here we go. Notice that imgui creates an ini file in the current
directory with position/size of the last dialogs used and it's
automatically reloaded the next time... !

There is a toggle in the cmake file to disable building with imgui but
it's untested, it should work though, it just disables all the new code
and re-enables the old campaign mission dialog.