sort of revert the hack for the black lines
it was due to the absence of the updat*mix files because I hadn't them
on the dos cds which I used to make the working directory.
Apparently these files contain the needed palette for the interpolation.
I added some comment about it in winasm.cpp in case it happens to
someone else !