comment out the mantissa precision field for now
it produces a ton of warnings when assigning -1 to it. Now I am not sure
how this precision is used, so a better solution might be to suppress
the warning. But for now I couldn't find any problem by just commenting
out this, it's commented in case it should be kept for some unknown
reason.