handle patch 1.6c translations
This allows to switch mix files based on a Language setting added to the
ini. I just added english, german and french, for japenese I am not
sure!
The briefings are not translated because the texts are either in the ini
for the mission, or in mission.ini for the campaign, the briefings from
the campaign are translated only if you use the right langage for
general.mix.
What is translated : icons, texts, speech and sounds.
Now it would be cool to be able to change the language directly from the
gui... maybe later !